Serve ui out of local html files; pass text/plain and image/* data to the webview using a data url
