+++ android/app/src/main/assets/index.html
@@ -1,0 +1,188 @@
+
+  <!DOCTYPE html>
+  <html lang="en-US">
+  <head>
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta charset="utf-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+
+    <meta property="og:title" content="Firefox Send"/>
+    <meta name="twitter:title" content="Firefox Send"/>
+    <meta name="description" content="Encrypt and send files with a link that automatically expires to ensure your important documents don’t stay online forever."/>
+    <meta property="og:description" content="Encrypt and send files with a link that automatically expires to ensure your important documents don’t stay online forever."/>
+    <meta name="twitter:description" content="Encrypt and send files with a link that automatically expires to ensure your important documents don’t stay online forever."/>
+    <meta name="twitter:card" content="summary"/>
+    <meta property="og:image" content="https://send.firefox.com/send-fb.19274ff0.jpg"/>
+    <meta name="twitter:image" content="https://send.firefox.com/send-twitter.7d5e4200.jpg"/>
+    <meta property="og:url" content="https://send.firefox.com"/>
+    <base href="https://send.firefox.com/" />
+
+    <title>Firefox Send</title>
+
+    <link rel="stylesheet" type="text/css" href="/style.11428181.css" />
+
+    <!-- generic favicons -->
+    <link rel="icon" href="/favicon-32.4efdbe5a.png" sizes="32x32">
+    <link rel="icon" href="/favicon-96.a2ce7c6e.png" sizes="96x96">
+    <link rel="icon" href="/favicon-128.58447464.png" sizes="128x128">
+    <link rel="icon" href="/favicon-228.069d14fe.png" sizes="228x228">
+
+    <!-- Android -->
+    <link rel="shortcut icon" href="/favicon-196.55cef95d.png" sizes="196x196">
+
+    <!-- iOS -->
+    <link rel="apple-touch-icon" href="/favicon-120.a48a3fc0.png" sizes="120x120">
+    <link rel="apple-touch-icon" href="/favicon-152.10d8b941.png" sizes="152x152">
+    <link rel="apple-touch-icon" href="/favicon-180.373787f8.png" sizes="180x180">
+
+    <!-- Windows 8 IE 10-->
+    <meta name="msapplication-TileColor" content="#FFFFFF">
+    <meta name="msapplication-TileImage" content="/favicon-144.909cb064.png">
+
+    <!-- Windows 8.1 + IE11 and above -->
+    <meta name="msapplication-config" content="/browserconfig.xml"/>
+
+
+    
+    <script defer src="/jsconfig.js"></script>
+    <script defer src="/runtime.44a28e7e.js"></script>
+    <script defer src="/vendor.ec5b5b1d.js"></script>
+    <script defer src="/public/locales/en-US/send.6b4f8354.js"></script>
+    <script defer src="/cryptofill.1315ac9e.js"></script>
+    <script defer src="/app.baa60a46.js"></script>
+  </head>
+  <body>
+      
+      
+  <header class="header">
+    <div class="logo">
+      <a class="logo__link" href="/">
+        <img
+          src="/send_logo.5fcfdf0e.svg"
+          alt="Send"/>
+        <h1 class="logo__title">Send</h1>
+      </a>
+      <div class="logo__subtitle">
+        <a class="logo__subtitle-link" href="https://testpilot.firefox.com">Firefox Test Pilot</a>
+        <div>web experiment</div>
+      </div>
+    </div>
+    <a href="https://qsurvey.mozilla.com/s3/txp-firefox-send?ver=2.5.4&amp;browser=unknown"
+      rel="noreferrer noopener"
+      class="feedback"
+      target="_blank">Feedback</a>
+  </header>
+      <main class="main">
+        <noscript>
+          <div class="noscript">
+            <h2>Firefox Send requires JavaScript</h2>
+            <p>
+              <a class="link" href="https://github.com/mozilla/send/blob/master/docs/faq.md#why-does-firefox-send-require-javascript">
+              Why does Firefox Send require JavaScript?
+              </a>
+            </p>
+            <p>Please enable JavaScript and try again.</p>
+          </div>
+        </noscript>
+        
+  <div id="page-one" class="">
+    <div class="title">Private, Encrypted File Sharing</div>
+    <div class="description">
+      <div>Send files through a safe, private, and encrypted link that automatically expires to ensure your stuff does not remain online forever.</div>
+      <a
+        href="https://testpilot.firefox.com/experiments/send"
+        class="link">
+        Learn more
+      </a>
+    </div>
+    <div class="uploadArea"
+     
+     >
+      <img
+        src="/upload.8e2a8bdb.svg"
+        title="Upload"/>
+      <div class="uploadArea__msg">
+        Drop your file here to start uploading
+      </div>
+      <span class="uploadArea__sizeMsg">
+        For the most reliable operation, it’s best to keep your file under 1GB
+      </span>
+      <input id="file-upload"
+        class="inputFile"
+        type="file"
+        name="fileUploaded"
+       
+       
+        />
+      <label for="file-upload"
+        class="btn btn--file"
+        title="Select a file to upload">
+        Select a file to upload
+      </label>
+    </div>
+    
+  </div>
+  
+      </main>
+      <footer class="footer">
+    <div class="legalSection">
+      <a
+        href="https://www.mozilla.org"
+        class="legalSection__link"
+        role="presentation">
+        <img
+          class="legalSection__mozLogo"
+          src="/mozilla-logo.2538e7d3.svg"
+          alt="mozilla"/>
+      </a>
+      <a
+        href="https://www.mozilla.org/about/legal"
+        class="legalSection__link">
+        Legal
+      </a>
+      <a
+        href="https://testpilot.firefox.com/about"
+        class="legalSection__link">
+        About Test Pilot
+      </a>
+      <a
+        href="/legal"
+        class="legalSection__link">Privacy</a>
+      <a
+        href="/legal"
+        class="legalSection__link">Terms</a>
+      <a
+        href="https://www.mozilla.org/privacy/websites/#cookies"
+        class="legalSection__link">
+        Cookies
+      </a>
+      <a
+        href="https://www.mozilla.org/about/legal/report-infringement/"
+        class="legalSection__link">
+        Report IP Infringement
+      </a>
+    </div>
+    <div class="socialSection">
+      <a
+        href="https://github.com/mozilla/send"
+        class="socialSection__link"
+        role="presentation">
+        <img
+          class="socialSection__icon"
+          src="/github-icon.74dc24f7.svg"
+          alt="github"/>
+      </a>
+      <a
+        href="https://twitter.com/FxTestPilot"
+        class="socialSection__link"
+        role="presentation">
+        <img
+          class="socialSection__icon"
+          src="/twitter-icon.018c7f54.svg"
+          alt="twitter"/>
+      </a>
+    </div>
+  </footer>
+    </body>
+  </html>
+
