+++ android/app/src/main/assets/intent-target.html
@@ -1,0 +1,20 @@
+
+  <!DOCTYPE html>
+  <html lang="en-US">
+  <head>
+
+    <title>Firefox Send</title>
+
+  </head>
+  <body>
+    <div id="output">output</div>
+<script>
+window.addEventListener("message", (event) => {
+  fetch(event.data).then(res => res.text()).then(txt => {
+    document.getElementById('output').textContent = "GOT MESSAGE" + txt;
+  });
+}, false);
+</script>
+  </body>
+  </html>
+
